# Speed up CLI tests